genlb: connection load balancer using ptrace
genlb intercepts and re-writes calls to connect(2) by ptrace'ing
subprocesses. The goal is a light weight service mesh for services
running in a container.

genlb works by:

* intercepting connect()'s by the child process
* reading out the sockaddr
* performing the same connect()
* using the LD_PRELOAD version of glb to intercept the connect()
* retrieving the peer address
* overwriting the sockaddr of the connect() call in the child process

The code is currently working on ARM and x86_64 platforms and supports
forked and threaded subprocesses. Currently only IPv4 and IPv4-mapped
IPv6 sockets are supported.

Example:

    # shell 1
    nc -vvvv -k -l 127.1.0.1:8080

    # shell 2
    nc -vvvv -k -l 127.2.0.1:8080

    # shell 3
    LD_PRELOAD=./libglb.so \
    GLB_OPTIONS="--random 127.0.0.1:8989 127.1.0.1:8080 127.2.0.1:8080" \
    ./genlb nc 127.0.0.1 8989

genlb will also resend some signals.

Signals are trapped in the child and forwarded to the parent. The
default behaviour of ptrace() causes the child to ignore the signal
when continued.

For example on ARM, executables run some tests at initialization
to determine if various instructions exist. Unsupported instructions
generate a SIGILL which the process catches.